requirements.txt
Django==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
celery==5.3.0
redis==5.0.0
groq==0.4.1
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
